requirements.txt
pandas
torch==1.11.0
transformers
python-dotenv
sendgrid